added numeric to schema.sql in place of INT to prevent int overflow

+++ schema.sql
@@ -3,6 +3,6 @@ DROP TABLE searches;
 CREATE TABLE searches(
   id SERIAL PRIMARY KEY,
   keyword VARCHAR(255),
-  google_results_count INT,
-  niche_score INT
+  google_results_count NUMERIC,
+  niche_score NUMERIC
 )
